move context types to own file to avoid circular deps
